initial post-processing framework
began implementing post-processing techniques such as tone mapping and bloom. changed some pipelines and removed some renders for better analysis of post-proc performance. also added more advanced fps stats to this end, and a way to /kind of/ gauge whether fps is being limited by refresh rate ("theoretical framerate" in troubleshooting text). other cleanup and suchwhat.

+++ src/PhysicsHandler.cpp
@@ -233,10 +233,9 @@ void PhysicsHandler::updateCameraPos (const std::vector<Tri>&land){
 }
 
 void PhysicsHandler::update (const std::vector<Tri>&land){
-	dt = std::chrono::duration<double>(std::chrono::high_resolution_clock::now() - lasttime).count();
+	dt = std::chrono::duration<float>(std::chrono::high_resolution_clock::now() - lasttime).count();
 	lasttime = std::chrono::high_resolution_clock::now();
 	t = std::chrono::duration<float>(lasttime - t0).count();
-	//should the lasttime reset go here???
 	updateCameraPos(land);
 }
